Split off Redash data formatting into its own module
